+++ rs/registry/canister-client/src/client/tests.rs
